backend/app/utils/auth.py
from fastapi import HTTPException, Security
from fastapi.security import HTTPBearer, HTTPAuthorizationCredentials
from jose import jwt, JWTError
from typing import Dict, Any
from app.models.user import User
import json
import base64

security = HTTPBearer()


async def get_current_user(credentials: HTTPAuthorizationCredentials = Security(security)):
    """Verify JWT token and return user info"""
    token = credentials.credentials
    
    try:
        # Try to decode using python-jose first
        try:
            decoded_token: Dict[str, Any] = jwt.decode(
                token,
                options={
                    "verify_signature": False,
                    "verify_exp": False,
                    "verify_nbf": False,
                    "verify_iat": False,
                    "verify_aud": False,
                    "verify_iss": False
                }
            )
        except Exception:
            # If jwt.decode fails, manually decode the payload
            # JWT format: header.payload.signature
            parts = token.split('.')
            if len(parts) != 3:
                raise HTTPException(status_code=401, detail="Invalid token format")
            
            # Decode the payload (second part)
            payload = parts[1]
            # Add padding if needed
            padding = 4 - len(payload) % 4
            if padding != 4:
                payload += '=' * padding
            decoded_bytes = base64.urlsafe_b64decode(payload)
            decoded_token = json.loads(decoded_bytes.decode('utf-8'))
        
        # Extract user information from token
        user_id = decoded_token.get("sub")
        email = decoded_token.get("email")
        
        if not user_id:
            raise HTTPException(status_code=401, detail="Invalid authentication token - no user ID")
        
        return User(id=user_id, email=email)
    except HTTPException:
        raise
    except Exception as e:
        raise HTTPException(status_code=401, detail=f"Authentication error: {str(e)}")
